refactor: interfaceからtypeへの統一的な変換
- すべてのinterfaceをtypeに変換してコードベースの一貫性を向上
- LikeButtonでAuthContextDynamicへの依存を更新
- 各hooksファイルでAuthContextDynamicへの依存を更新
- 型定義のスタイルを統一してメンテナンス性を向上

変更ファイル:
- components/like-button.tsx
- components/global-search/hooks/use-global-search.ts
- hooks/use-like.ts
- hooks/use-meshi-like-state.ts
- hooks/use-my-likes.ts
- hooks/use-my-profile.ts
