+++ apps/frontend/hooks/use-meshi-like-state.ts
@@ -2,7 +2,7 @@
 
 import { request } from 'graphql-request'
 import { useEffect, useState } from 'react'
-import { useAuth } from '@/contexts/AuthContext'
+import { useAuth } from '@/contexts/AuthContextDynamic'
 
 const MeshiLikeStateQuery = `
   query MeshiLikeState($meshiId: ID!) {
@@ -14,7 +14,7 @@ const MeshiLikeStateQuery = `
   }
 `
 
-interface UseMeshiLikeStateReturn {
+type UseMeshiLikeStateReturn = {
   isLiked: boolean
   likeCount: number
   loading: boolean
